% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess4_dataframe.R
\docType{methods}
\name{as.data.frame,ess4_dataframe-method}
\alias{as.data.frame,ess4_dataframe-method}
\title{convert ess4_dataframe back to data.frame}
\usage{
\S4method{as.data.frame}{ess4_dataframe}(x, row.names = NULL,
  option = FALSE, ..., stringsAsFactors = FALSE)
}
\arguments{
\item{x}{an ess4_dataframe}

\item{row.names}{defaults to NULL}

\item{option}{boolean defaults to FALSE}

\item{...}{all the other good stuff}

\item{stringsAsFactors}{boolean defaults to FALSE because why would you want to do anything else? note that this is not the default behavior of an as.data.frame method.}
}
\value{
a \code{\link{dataframe}}
}
\description{
i use this to quickly grab the underlying data.frame mainly in order to pass it cleanly to S3 methods, but there are other applications.
}
\examples{
df1 	<- data.frame(a = rnorm(10), b = rnorm(10))
dfx	<- ess4_dataframe(df1)
df2 	<- as.data.frame(dfx)
identical(df1, df2)
}
\seealso{
\code{\link{ess4_dataframe}}
}
