% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_load_tags.R
\name{batch_load_tags}
\alias{batch_load_tags}
\title{batch load sat tags}
\usage{
batch_load_tags(data_dir, tags_dir = NA, streams = NA)
}
\arguments{
\item{data_dir}{character points to a directory of sat tags}

\item{tags_dir}{character vector of tag directory names to include or NA defaults to every directory in \code{data_dir}}

\item{streams}{character vector limiting which streams to search for. NA defaults to all streams. note \code{*-Summary.csv} is expected to populate some of the slots of \code{sattag}.}
}
\value{
a list of \code{\link[sattagutils]{sattag-class}} S4 objects.
}
\description{
load sat tags from a directory containing directories downloaded from the portal
}
\examples{
\dontrun{
tags <- load_tag("~/path/to/tags/")
}
}
\seealso{
Other tag stream loaders: \code{\link{load_tag}}
}
\concept{tag stream loaders}
