% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_beh.R
\name{censor_beh}
\alias{censor_beh}
\title{censor a behavior stream}
\usage{
censor_beh(b1, depth = 50, duration = 33 * 60)
}
\arguments{
\item{b1}{a behavior data stream or a dataframe approximating one. requires columns \code{DurationMin}, \code{DurationMax}, \code{What}, \code{Start}, \code{End}, \code{DepthMax}, \code{DepthMin}.}

\item{depth}{the minimum depth required to qualify as a dive}

\item{duration}{the minimum duration required to qualify as a dive in seconds.}
}
\description{
based on depth and duration qualifications
}
