% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateseq.R
\name{dateseq}
\alias{dateseq}
\title{quickly create nice data sequences for plotting}
\usage{
dateseq(d, hours = FALSE)
}
\arguments{
\item{d}{a numeric vector of times}

\item{hours}{boolean defaults to FALSE. when TRUE will return hourly times.}
}
\description{
using numeric times (seconds since UNIX epoch) find where to put the daily and hourly tickmarks
}
\details{
probably could make this more general, but don't need to right now.
}
\examples{
dateseq(1:(5*24*60*60))
}
