% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R, R/tagstack.R
\docType{methods}
\name{DeployID}
\alias{DeployID}
\alias{DeployID,sattag-method}
\alias{DeployID,tagstack-method}
\title{get DeployID}
\usage{
DeployID(x)

\S4method{DeployID}{sattag}(x)

\S4method{DeployID}{tagstack}(x)
}
\description{
function to extract DeployID.
}
\section{Methods (by class)}{
\itemize{
\item \code{sattag}: method for sattag

\item \code{tagstack}: method for tagstack
}}

\seealso{
Other slot access functions: \code{\link{Ptt}},
  \code{\link{filename}}, \code{\link{instrument}},
  \code{\link{loadtime}}, \code{\link{location}},
  \code{\link{species}}, \code{\link{streamtype}},
  \code{\link{tagdir}}, \code{\link{tagstackdir}}
}
\concept{slot access functions}
