% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess4_dataframe.R
\name{ess4_dataframe}
\alias{ess4_dataframe}
\title{ess4_dataframe constructor}
\usage{
ess4_dataframe(..., stringsAsFactors = FALSE)
}
\arguments{
\item{...}{all that good regular data.frame stuff}

\item{stringsAsFactors}{boolean defaults to FALSE because why would you do anything else?}
}
\value{
an S4 class of \code{ess4_dataframe}.
}
\description{
this would be the typical way to instantiate an ess4_dataframe.
}
\examples{
ess4_dataframe(a = rnorm(10), b = rnorm(10))
}
\seealso{
\code{\link[sattagutils]{ess4_dataframe-class}}
}
