% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream.R
\docType{methods}
\name{filename}
\alias{filename}
\alias{filename,sattagstream-method}
\title{get filename of a sattagstream}
\usage{
filename(x)

\S4method{filename}{sattagstream}(x)
}
\arguments{
\item{x}{source filename for a csv data stream file.}
}
\description{
use this function to get the filename of a \code{\link[sattagutils]{sattagstream-class}}. changing the source filename after construction is currently not supported.
}
\section{Methods (by class)}{
\itemize{
\item \code{sattagstream}: get the source filename of a sattagstream
}}

\seealso{
\code{\link[sattagutils]{sattagstream}}
}
