% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstream.R
\name{get.stream}
\alias{get.stream}
\title{extract a stream from a sattag.}
\usage{
get.stream(tag, stream)
}
\arguments{
\item{tag}{an S4 object which extends \code{\link[sattagutils]{sattag}}.}

\item{either}{a character (stream name) or numeric (stream index).}
}
\value{
either a list of S4 \code{\link[sattagutils]{sattagstream}} objects or a single S4 sattagstream if \code{length(stream) == 1}.
}
\description{
perferred method for extracting a sat tag stream.
}
\examples{
\dontrun{
tag1 <- load_tag("path/to/tags/tag1")
getStream(tag1, 1)
	getStream(tag1, 'argos')	
}
}
\concept{sattagstream manipulators}
