% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R, R/tagstack.R
\docType{methods}
\name{getstream}
\alias{getstream}
\alias{getstream,sattag-method}
\alias{getstream,tagstack-method}
\title{get a stream}
\usage{
getstream(x, type, squash = FALSE)

\S4method{getstream}{sattag}(x, type, squash = FALSE)

\S4method{getstream}{tagstack}(x, type, squash = FALSE)
}
\arguments{
\item{type}{streamtype to extract}

\item{squash}{boolean defaults to FALSE. If TRUE, will return a single sattagstream.}
}
\value{
either a \code{tagstack} or \code{sattag} depending on input.
}
\description{
function to get a particular streamtype from a \code{\link[sattagutils]{sattag}} or \code{\link[sattagutils]{tagstack}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{sattag}: return all streams of streamtype type

\item \code{tagstack}: return a tagstack of all streams of type
}}

\seealso{
Other slot access functions: \code{\link{DeployID}},
  \code{\link{Ptt}}, \code{\link{filename}},
  \code{\link{instrument}}, \code{\link{loadtime}},
  \code{\link{location}}, \code{\link{species}},
  \code{\link{streamtype}}, \code{\link{tag_en}},
  \code{\link{tag_st}}, \code{\link{tagdir}},
  \code{\link{tagstackdir}}
}
\concept{slot access functions}
