% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_stream.R
\name{has_stream}
\alias{has_stream}
\title{has_stream}
\usage{
has_stream(x, streamname)
}
\arguments{
\item{x}{the tagstack or sattag in question}

\item{streamname}{the stream you are searching for}
}
\value{
a bool if any of the streams are \code{streamname}
}
\description{
Checks to see if a tagstack or sattag has a stream
}
