% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag_init.R
\docType{methods}
\name{initialize,sattag-method}
\alias{initialize,sattag-method}
\title{initialize an S4 object of class \code{\link[sattagutils]{sattag}}}
\usage{
\S4method{initialize}{sattag}(.Object, ...)
}
\description{
initialize an S4 object of class \code{\link[sattagutils]{sattag}}
}
