% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isfunctions.R
\name{isfunctions}
\alias{isfunctions}
\alias{is.sattag}
\alias{is.sattagstream}
\title{is functions for sattag and sattagstream}
\usage{
is.sattag(obj)

is.sattagstream(obj)
}
\description{
check to see if an object is a \code{\link[sattagutils]{sattag}} or \code{\link[sattagutils]{sattagstream}}
}
\examples{
\dontrun{
tag1 <- load_tag("path/to/tags/tag1")
s1 <- getStream(tag1, 1)

c(is.stream(s1), is.tag(s1), is.sattag(tag1), is.stream(tag1))
}
}
