% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlond.hav.R
\name{latlond.hav}
\alias{latlond.hav}
\title{calculate the haversine distance between two geographic coordiantes}
\usage{
latlond.hav(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1, lon1}{start coordinates (can be vectors)}

\item{lat2, lon2}{end coordinates (can be vectors)}
}
\value{
distance in kilometers
}
\description{
calculate the haversine distance between two geographic coordiantes
}
\details{
uses 6371 as the radius of the earth in kilometers.
}
\examples{
latlond.hav(-6.72, 147, -4.67, -174.52)
}
\references{
\url{https://www.movable-type.co.uk/scripts/latlong.html}
}
\seealso{
Other distance functions: 
\code{\link{torad}()}
}
\concept{distance functions}
