% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tag.R
\name{load_tag}
\alias{load_tag}
\title{load a single sat tag}
\usage{
load_tag(tag_dir, streams = NA)
}
\arguments{
\item{tag_dir}{a path to a tag directory containg csv data streams downloaded from the portal.}

\item{streams}{a character vector limiting which streams to search for. NA default to all streams. note that \code{*-Summary.csv} is expected to populate some of the slots.}
}
\value{
a \code{\link[sattagutils]{sattag}} S4 object.
}
\description{
given a tag directory from a portal download instantiate, populate, and return a \code{\link[sattagutils]{sattag}} S4 object.
}
\examples{
\dontrun{
tag1 <- load_tag("~/path/to/tags/tag1")
}
}
\seealso{
Other tag stream loaders: \code{\link{batch_load_tags}}
}
\concept{tag stream loaders}
