% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R, R/tagstack.R
\docType{methods}
\name{loadtime}
\alias{loadtime}
\alias{loadtime,sattag-method}
\alias{loadtime,tagstack-method}
\title{get loadtime}
\usage{
loadtime(x)

\S4method{loadtime}{sattag}(x)

\S4method{loadtime}{tagstack}(x)
}
\description{
function to extract time a tag was loaded into R originally based on when \code{intialize()} was called.
}
\section{Methods (by class)}{
\itemize{
\item \code{sattag}: method for sattag

\item \code{tagstack}: method for tagstack
}}

\seealso{
Other slot access functions: \code{\link{DeployID}},
  \code{\link{Ptt}}, \code{\link{filename}},
  \code{\link{getstream}}, \code{\link{instrument}},
  \code{\link{location}}, \code{\link{species}},
  \code{\link{streamtype}}, \code{\link{tag_en}},
  \code{\link{tag_st}}, \code{\link{tagdir}},
  \code{\link{tagstackdir}}
}
\concept{slot access functions}
