% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R, R/tagstack.R
\docType{methods}
\name{location}
\alias{location}
\alias{location,sattag-method}
\alias{location,tagstack-method}
\title{get location}
\usage{
location(x)

\S4method{location}{sattag}(x)

\S4method{location}{tagstack}(x)
}
\description{
function to extract study location as defined in *-Labels.csv.
}
\section{Methods (by class)}{
\itemize{
\item \code{sattag}: method for sattag

\item \code{tagstack}: method for tagstack
}}

\seealso{
Other slot access functions: \code{\link{DeployID}},
  \code{\link{Ptt}}, \code{\link{filename}},
  \code{\link{getstream}}, \code{\link{instrument}},
  \code{\link{loadtime}}, \code{\link{species}},
  \code{\link{streamtype}}, \code{\link{tag_en}},
  \code{\link{tag_st}}, \code{\link{tagdir}},
  \code{\link{tagstackdir}}
}
\concept{slot access functions}
