% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchtimes.R
\name{matchtimes}
\alias{matchtimes}
\title{a tiny time matching function}
\usage{
matchtimes(t1, t2)
}
\arguments{
\item{t1, t2}{numeric time vectors.}
}
\description{
see \code{\link[base]{findInterval}}.
}
\examples{
t1 <- c(10, 20, 30, 40)
t2 <- c(9, 21, 36, 37)
matchtimes(t1, t2)
matchtimes(t2, t1)
}
