% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es4dataframe.R
\name{merge,es4dataframe-method}
\alias{merge,es4dataframe-method}
\title{wrapper for S3 method}
\usage{
\S4method{merge}{es4dataframe}(
  x,
  y,
  by = intersect(names(x), names(y)),
  by.x = by,
  by.y = by,
  all = FALSE,
  all.x = all,
  all.y = all,
  sort = TRUE,
  suffixes = c(".x", ".y"),
  no.dups = TRUE,
  incomparables = NULL,
  ...
)
}
\description{
wrapper for S3 method
}
