% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_stacks.R
\name{merge_stacks}
\alias{merge_stacks}
\title{merge tagstacks}
\usage{
merge_stacks(target_stack, source_stack, by = "Ptt",
  remove_duplicates = FALSE, identify_original = FALSE,
  target_lab = "target", source_lab = "source")
}
\arguments{
\item{by}{the name of the \code{sattag} slot to base the merge on. Should be \code{"Ptt"} or \code{"DeployID"}. Defaults to \code{"Ptt"}.}

\item{remove_duplicates}{a boolean defaults to FALSE. If TRUE, duplicated rows of data will be retained in the output tagstack.}

\item{identify_original}{a boolean defaults to FALSE. If TRUE, a character 'original' column will be added to each sattag stream indicating the origin of each row with a character identifier specified by target_lab or source_lab.}

\item{target_lab}{a string to identify data that came from the target tagstack.}

\item{source_lab}{a string to identify data that came from the source tagstack.}

\item{target}{a tagstack, the target stack}

\item{source}{a tagstack, the source stack}
}
\value{
an S4 object of class \code{\link[sattagutils]{tagstack}}.
}
\description{
sensibily merge two tagstacks together with overlapping sets of tags
}
