% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2date.R
\name{num2date}
\alias{num2date}
\title{translate numeric dates back to \code{POSIX*}}
\usage{
num2date(d, tz = "UTC", origin = "1970-01-01", ...)
}
\arguments{
\item{d}{a numeric vector of dates.}

\item{tz, origin}{as expected by \code{\link[base]{as.POSIXct}}.}
}
\description{
will convert a numeric vector (seconds since UNIX epoch) using \code{as.POSIXct}. opposite of \code{\link[sattagutils]{date2num}}. methods for \code{sattagstream} classes not implemeneted.
}
\examples{
num2date(574153200)
}
