% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate_series.R
\name{paginate_series}
\alias{paginate_series}
\title{a pagination tool for plotting series datastream.}
\usage{
paginate_series(s, increment = 4 * 48, time = "samples", xaxt = "n",
  las = 1, ...)
}
\arguments{
\item{s}{a series data stream or a dataframe approximating one. gets passed to  \code{\link[sattagutils]{plot_series}}.}

\item{increment}{the number of sampling points to show on the plot at one time. also current increments without overlap when you page forward or backwards.}

\item{time}{a string which indicates what you want the x-axis to be. right now only "samples" (running sample index) or "numeric" (matlab style datenum since the unix epoch) are supported.}

\item{xaxt}{default set to 'n' so i can plot an axis like you asked for in time.}

\item{las}{default to 1 why would you do anything else?}

\item{...}{all that other good graphics stuff and gets passed to \code{\link[sattagutils]{plot_series}}.}
}
\description{
simple function which provides a menu navigation for seires datastream
}
