% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series.R
\name{plot_series}
\alias{plot_series}
\title{plotting function for series datastream.}
\usage{
plot_series(s, show_gaps = FALSE, new = TRUE, points = TRUE,
  lines = TRUE, pch = 16, cex = 0.5, col.lines = "black",
  col.points = "black", ...)
}
\arguments{
\item{s}{a series data stream or a dataframe approximating one. requires columns numeric \code{Date}, numeric \code{Depth}.right now also requires to have at least 2 rows that is so i can calculate the sampling period. see readme for more information about this.}

\item{show_gaps}{boolean defaults to FALSE. if TRUE pink rectangles will show the data gaps}

\item{new}{boolean defaults to TRUE. draw a new plot if this is TRUE otherwise draw on whatever you have up.}

\item{points}{boolean defaults to TRUE. draw points.}

\item{lines}{boolean defaults to TRUE. draw lines.}

\item{pch, cex}{just reasonable defaults do whatever you want.}

\item{col.lines}{color of the lines if drawn.}

\item{col.points}{color of the points if drawn.}

\item{...}{all the other good plotting stuff. gets passed to \code{plot}, \code{points}, and \code{lines} so think about it.}
}
\description{
simple function to plot series data quickly.
}
