% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R, R/tagstack.R
\docType{methods}
\name{Ptt}
\alias{Ptt}
\alias{Ptt,sattag-method}
\alias{Ptt,tagstack-method}
\title{get Ptt}
\usage{
Ptt(x)

\S4method{Ptt}{sattag}(x)

\S4method{Ptt}{tagstack}(x)
}
\value{
a character representation of the Ptt. why this and not numeric?
}
\description{
function to extract Ptt.
}
\section{Methods (by class)}{
\itemize{
\item \code{sattag}: method for sattag

\item \code{tagstack}: method for tagstack
}}

\seealso{
Other slot access functions: \code{\link{DeployID}},
  \code{\link{filename}}, \code{\link{getstream}},
  \code{\link{instrument}}, \code{\link{loadtime}},
  \code{\link{location}}, \code{\link{species}},
  \code{\link{streamtype}}, \code{\link{tag_en}},
  \code{\link{tag_st}}, \code{\link{tagdir}},
  \code{\link{tagstackdir}}
}
\concept{slot access functions}
