% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream_extends.R
\docType{class}
\name{rawargos-class}
\alias{rawargos-class}
\title{an S4 class which extends sattagstream}
\description{
corresponds to the *-RawArgos.csv data stream. this is not a well formated csv when downloaded from the portal so beware when importing. see \code{\link[sattagutils]{load_tag}}.
}
\seealso{
\code{\link[sattagutils]{sattag-class}}

\code{\link[sattagutils]{sattagstream-class}}

Other sattagstream types: 
\code{\link{all-class}},
\code{\link{argos-class}},
\code{\link{behavior-class}},
\code{\link{corrupt-class}},
\code{\link{fastgps-class}},
\code{\link{histos-class}},
\code{\link{labels-class}},
\code{\link{locations-class}},
\code{\link{minmaxdepth-class}},
\code{\link{rtc-class}},
\code{\link{series-class}},
\code{\link{seriesrange-class}},
\code{\link{sst-class}},
\code{\link{status-class}},
\code{\link{summary-class}}
}
\concept{sattagstream types}
