% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcsv.R
\name{rcsv}
\alias{rcsv}
\title{sensible defaults for \code{read.table}}
\usage{
rcsv(..., header = TRUE, sep = ",", stringsAsFactors = FALSE)
}
\arguments{
\item{...}{all the regular good stuff for finding files and whatnot.}

\item{header}{i always want a header}

\item{sep}{we're talking about comma seperated values here.}

\item{stringsAsFactors}{default is FALSE. please don't change that. why would you change that?}
}
\description{
sensible defaults for \code{read.table}
}
\details{
please don't use read_csv or any of that other nonesense, it breaks everyone's code...
}
\examples{
# use it just like read.table but without the worry
\dontrun{
rcsv("file.csv")
}
}
\seealso{
Other sensible csv functions: 
\code{\link{wcsv}()}
}
\concept{sensible csv functions}
