% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_ser.R
\name{resample_ser}
\alias{resample_ser}
\title{resample series}
\usage{
resample_ser(s, sampling_period)
}
\arguments{
\item{s}{a series data stream or a dataframe approximating one. requires column numeric \code{Date}.}

\item{sampling_period}{desired sampling period in minutes (sorry).}
}
\description{
downsample series to a particular sampling period.
}
