% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R
\docType{class}
\name{sattag-class}
\alias{sattag-class}
\title{an S4 class to represent a single sat tag}
\description{
represents a single sat tag. holds a small amount of meta data and a list of data streams (extends \code{\link{list}}).
}
\section{Slots}{

\describe{
\item{\code{instrument}}{instrument type, e.g., "MK10-A"}

\item{\code{DeployID}}{character vector}

\item{\code{Ptt}}{character vector}

\item{\code{species}}{character vector}

\item{\code{location}}{character vector of the study site name.}

\item{\code{earliestdata,latestdata}}{numeric dates (seconds since UNIX epoch).}

\item{\code{directory}}{the original directory from which the tag was loaded.}

\item{\code{loadtime}}{a character vector set by \code{initialize()} when object is instantiated.}
}}

