% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R
\name{sattag}
\alias{sattag}
\title{constructor for sattag}
\usage{
sattag(
  data = list(),
  instrument = character(),
  DeployID = character(),
  Ptt = character(),
  species = character(),
  location = character(),
  t_start = numeric(),
  t_end = numeric(),
  directory = character()
)
}
\arguments{
\item{data}{this should be a list of \code{\link[sattagutils]{sattagstreams-class}}.}

\item{instrument}{instrument type, e.g., "MK10-A"}

\item{DeployID}{character vector}

\item{Ptt}{character vector}

\item{species}{character vector}

\item{location}{character vector of the study site name.}

\item{t_start, t_end}{numeric dates (seconds since UNIX epoch).}

\item{directory}{the original directory from which the tag was loaded.}
}
\value{
a \code{\link[sattagutils]{sattag-class}}
}
\description{
use this constructor to create a new sattag object.
}
