% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream.R
\docType{class}
\name{sattagstream}
\alias{sattagstream}
\title{an S4 class to represent a single data stream in a tag}
\description{
a parent class for specific sattag stream types. generally accessed from within a \code{\link[sattagutils]{sattag}}.
}
\section{Slots}{

\describe{
\item{\code{streamname}}{character identifer for the stream.}

\item{\code{filename}}{the original filename from which the data stream was derrived. this will almost always be a text or csv file.}

\item{\code{data}}{data frame representing the actual data stream. almost all wildlife computer data streams (csvs) are in tabular format so this makes sense, but in the future perhaps this should be made more general (e.g., list).}
}}

\seealso{
Other sattagstream types: \code{\link{stream_all}},
  \code{\link{stream_argos}},
  \code{\link{stream_behavior}},
  \code{\link{stream_corrupt}},
  \code{\link{stream_fastgps}}, \code{\link{stream_hists}},
  \code{\link{stream_labels}},
  \code{\link{stream_locations}},
  \code{\link{stream_rawargos}}, \code{\link{stream_rtc}},
  \code{\link{stream_seriesrange}},
  \code{\link{stream_series}}, \code{\link{stream_sst}},
  \code{\link{stream_status}}, \code{\link{stream_summary}}
}
\concept{sattagstream types}
