% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ser2beh.R
\name{ser2beh}
\alias{ser2beh}
\title{downsample series data to faux behavior data}
\usage{
ser2beh(s, surface_threshold_meters = 25,
  vrate_ascent_meters_per_second = 0.7,
  vrate_decent_meters_per_second = 1.5,
  dive_definition_threshold_meters = 50, period)
}
\arguments{
\item{s}{a series data stream or a dataframe approximating one. requires columns numeric \code{Date}, numeric \code{Depth}.}

\item{dive_definition_threshold_meters}{would be from the behavior settings you want to emulate. this is the threshold to qualify as a behavior dive.}

\item{period}{sampling period of input series data}

\item{surface_threshold_metersused}{by the peak finding algorithm to determine if a peak is close enough to the surface to represent a real surfacing event. default 25 is for me. you will need ot pick something that makes sense for your species.}

\item{vrate_ascent_meters_per_second.}{this is the assumed mean vertical accent rate to interpolate surfacing times. default 0.7 is for Ziphius cavirostris (Tyack et al. 2006).}

\item{vrate_decent_meters_per_second.}{this is the assumed mean vertical decent rate to interpolate surfacing times. default 1.5 for Ziphius cavirostris (Tyack et al. 2006)}
}
\description{
a simple downsampling which takes series data stream as an input and creates an estimate of what the behavior stream would have looked like. tries to interpolate surfacing times for more accurate duration estimates.
}
\references{
Tyack, P. L., Johnson, M., Soto, N. A., Sturlese, A., & Madsen, P. T. (2006). Extreme diving of beaked whales. Journal of Experimental Biology, 209(21), 4238–4253. https://doi.org/10.1242/jeb.02505
}
