% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag_show.R
\docType{methods}
\name{show,sattag-method}
\alias{show,sattag-method}
\title{display an S4 object of class \code{sattag}}
\usage{
\S4method{show}{sattag}(object)
}
\description{
display an S4 object of class \code{sattag}
}
