% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by_message.R
\name{sort_by_message}
\alias{sort_by_message}
\title{sort behavior streams}
\usage{
sort_by_message(beh)
}
\arguments{
\item{beh}{a beavior sattagstream.}
}
\value{
a behavior sattagstream sorted.
}
\description{
sort behavior streams by start time keeping messages together
}
