% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R, R/tagstack.R
\docType{methods}
\name{species}
\alias{species}
\alias{species,sattag-method}
\alias{species,tagstack-method}
\title{get species}
\usage{
species(x)

\S4method{species}{sattag}(x)

\S4method{species}{tagstack}(x)
}
\description{
function to extract species as defined in *-Labels.csv.
}
\section{Methods (by class)}{
\itemize{
\item \code{sattag}: method for sat tag

\item \code{tagstack}: method for tagstack
}}

\seealso{
Other slot access functions: \code{\link{DeployID}},
  \code{\link{Ptt}}, \code{\link{filename}},
  \code{\link{instrument}}, \code{\link{loadtime}},
  \code{\link{location}}, \code{\link{streamtype}},
  \code{\link{tagdir}}, \code{\link{tagstackdir}}
}
\concept{slot access functions}
