% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream_extends.R
\docType{class}
\name{stream_fastgps-class}
\alias{stream_fastgps-class}
\title{an S4 class which extends \code{sattagstream}}
\description{
corresponds to the *-FastGPS.csv data stream. this is not a well formated csv when downloaded from the portal so beware when importing. see load_stream.
}
