% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream_extends.R
\docType{class}
\name{stream_labels-class}
\alias{stream_labels-class}
\title{an S4 class which extends \code{sattagstream}}
\description{
corresponds to the *-Labels.csv data stream. this csv file appears to be missing an EOF when downloaded from the portal so beware when importing. see load_stream.
}
