% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream_extends.R
\docType{class}
\name{stream_labels}
\alias{stream_labels}
\title{an S4 class which extends sattagstream}
\description{
corresponds to the *-Labels.csv data stream. this csv file appears to be missing an EOF when downloaded from the portal so beware when importing. see \code{\link[sattagutils]{load_tag}}.
}
\seealso{
\code{\link[sattagutils]{sattag}}

\code{\link[sattagutils]{sattagstream}}

Other sattagstream types: \code{\link{sattagstream}},
  \code{\link{stream_all}}, \code{\link{stream_argos}},
  \code{\link{stream_behavior}},
  \code{\link{stream_corrupt}},
  \code{\link{stream_fastgps}}, \code{\link{stream_hists}},
  \code{\link{stream_locations}},
  \code{\link{stream_rawargos}}, \code{\link{stream_rtc}},
  \code{\link{stream_seriesrange}},
  \code{\link{stream_series}}, \code{\link{stream_sst}},
  \code{\link{stream_status}}, \code{\link{stream_summary}}
}
\concept{sattagstream types}
