% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream_extends.R
\docType{class}
\name{stream_series-class}
\alias{stream_series-class}
\title{an S4 class which extends sattagstream}
\description{
corresponds to the *-Series.csv data stream.
}
\seealso{
\code{\link[sattagutils]{sattag-class}}

\code{\link[sattagutils]{sattagstream-class}}

Other sattagstream types: \code{\link{stream_all-class}},
  \code{\link{stream_argos-class}},
  \code{\link{stream_behavior-class}},
  \code{\link{stream_corrupt-class}},
  \code{\link{stream_fastgps-class}},
  \code{\link{stream_histos-class}},
  \code{\link{stream_labels-class}},
  \code{\link{stream_locations-class}},
  \code{\link{stream_minmaxdepth-class}},
  \code{\link{stream_rawargos-class}},
  \code{\link{stream_rtc-class}},
  \code{\link{stream_seriesrange-class}},
  \code{\link{stream_sst-class}},
  \code{\link{stream_status-class}},
  \code{\link{stream_summary-class}}
}
\concept{sattagstream types}
