% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream_extends.R
\docType{class}
\name{stream_seriesrange-class}
\alias{stream_seriesrange-class}
\title{an S4 class which extends \code{sattagstream}}
\description{
corresponds to the *-SeriesRange.csv data stream.
}
