% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream_extends.R
\docType{class}
\name{stream_summary}
\alias{stream_summary}
\title{an S4 class which extends sattagstream}
\description{
corresponds to the *-Summary.csv data stream. used to populate slots in \code{\link[sattagutils]{sattag}} within \code{\link[sattagutils]{load_tag}}.
}
\seealso{
\code{\link[sattagutils]{sattag}}

\code{\link[sattagutils]{sattagstream}}

Other sattagstream types: \code{\link{sattagstream}},
  \code{\link{stream_all}}, \code{\link{stream_argos}},
  \code{\link{stream_behavior}},
  \code{\link{stream_corrupt}},
  \code{\link{stream_fastgps}}, \code{\link{stream_hists}},
  \code{\link{stream_labels}},
  \code{\link{stream_locations}},
  \code{\link{stream_rawargos}}, \code{\link{stream_rtc}},
  \code{\link{stream_seriesrange}},
  \code{\link{stream_series}}, \code{\link{stream_sst}},
  \code{\link{stream_status}}
}
\concept{sattagstream types}
