% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es4dataframe.R
\docType{methods}
\name{[,es4dataframe-method}
\alias{[,es4dataframe-method}
\title{wrapper for S3 method}
\usage{
\S4method{[}{es4dataframe}(x, i, j, ..., drop = TRUE)
}
\description{
wrapper for S3 method
}
