% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R
\docType{methods}
\name{[,sattag-method}
\alias{[,sattag-method}
\title{definition for the subset operator [}
\usage{
\S4method{[}{sattag}(x, i, j, ..., drop = TRUE)
}
\description{
whae i want this to do is return a sattag as opposed to [[ which should return a stream (and it does by default).
}
