% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagstack.R
\docType{methods}
\name{[,tagstack-method}
\alias{[,tagstack-method}
\title{definition for the subset operator}
\usage{
\S4method{[}{tagstack}(x, i, j, ..., drop = TRUE)
}
\description{
what i want this to do is return a tagstack as opposed to [[ which should return a tag (and it does by default).
}
