% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream_extends.R
\docType{class}
\name{summary-class}
\alias{summary-class}
\title{an S4 class which extends sattagstream}
\description{
corresponds to the *-Summary.csv data stream. used to populate slots in \code{\link[sattagutils]{sattag}} within \code{\link[sattagutils]{load_tag}}.
}
\seealso{
\code{\link[sattagutils]{sattag-class}}

\code{\link[sattagutils]{sattagstream-class}}

Other sattagstream types: 
\code{\link{all-class}},
\code{\link{argos-class}},
\code{\link{behavior-class}},
\code{\link{corrupt-class}},
\code{\link{fastgps-class}},
\code{\link{histos-class}},
\code{\link{labels-class}},
\code{\link{locations-class}},
\code{\link{minmaxdepth-class}},
\code{\link{rawargos-class}},
\code{\link{rtc-class}},
\code{\link{series-class}},
\code{\link{seriesrange-class}},
\code{\link{sst-class}},
\code{\link{status-class}}
}
\concept{sattagstream types}
