% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R, R/tagstack.R
\docType{methods}
\name{tag_en}
\alias{tag_en}
\alias{tag_en,sattag-method}
\alias{tag_en<-,sattag-method}
\alias{tag_en,tagstack-method}
\title{get end time of tag}
\usage{
tag_en(x)

\S4method{tag_en}{sattag}(x)

\S4method{tag_en}{sattag}(x) <- value

\S4method{tag_en}{tagstack}(x)
}
\description{
function to extract the data end time for a particular tag. if \code{\link[sattagutils]{load_tag}} or \code{\link[sattagutils]{batch_load_tags}} created the tag object then this is set from the \code{LatestDataTime} in the \code{summary} stream.
}
\section{Methods (by class)}{
\itemize{
\item \code{sattag}: method for sattag

\item \code{sattag}: method for sattag

\item \code{tagstack}: method for tagstack
}}

\seealso{
Other slot access functions: \code{\link{DeployID}},
  \code{\link{Ptt}}, \code{\link{filename}},
  \code{\link{instrument}}, \code{\link{loadtime}},
  \code{\link{location}}, \code{\link{species}},
  \code{\link{streamtype}}, \code{\link{tag_st}},
  \code{\link{tagdir}}, \code{\link{tagstackdir}}
}
\concept{slot access functions}
