% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R
\name{tag_st<-}
\alias{tag_st<-}
\title{set start time of tag}
\usage{
tag_st(x) <- value
}
\description{
function to set the data end time for a particular tag. if \code{\link[sattagutils]{load_tag}} or \code{\link[sattagutils]{batch_load_tags}} created the tag object then this is initially set from the \code{EarliestDataTime} in the \code{summary} stream.
}
