% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R, R/tagstack.R
\docType{methods}
\name{tagdir}
\alias{tagdir}
\alias{tagdir,sattag-method}
\alias{tagdir,tagstack-method}
\title{get tagdir}
\usage{
tagdir(x)

\S4method{tagdir}{sattag}(x)

\S4method{tagdir}{tagstack}(x)
}
\description{
function to extract source directory from which the tag was loaded.
}
\section{Methods (by class)}{
\itemize{
\item \code{sattag}: method for sattag

\item \code{tagstack}: method for tagstack
}}

\seealso{
Other slot access functions: \code{\link{DeployID}},
  \code{\link{Ptt}}, \code{\link{filename}},
  \code{\link{instrument}}, \code{\link{loadtime}},
  \code{\link{location}}, \code{\link{species}},
  \code{\link{streamtype}}, \code{\link{tag_en}},
  \code{\link{tag_st}}, \code{\link{tagstackdir}}
}
\concept{slot access functions}
