% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagstack.R
\docType{class}
\name{tagstack-class}
\alias{tagstack-class}
\title{an S4 class to represent a stack of sat tags}
\description{
represents a tag stack. wrapper for a \code{\link{list}}.
}
\seealso{
\code{\link[sattagutils]{tagstack}}
}
