% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagstack.R
\name{tagstack}
\alias{tagstack}
\title{constructor for tagstack}
\usage{
tagstack(data = list(), directory = character())
}
\arguments{
\item{data}{this should be a list of \code{\link[sattagutils]{sattag-class}}.}
}
\value{
a \code{\link[sattagutils]{tagstack-class}}
}
\description{
use this constructor to create a new tagstack object.
}
