% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagstack.R
\name{tagstackdir}
\alias{tagstackdir}
\title{get tagstack source directory}
\usage{
tagstackdir(x)
}
\arguments{
\item{a}{\code{\link[sattagutils]{tagstack}}}
}
\value{
a character representation of the source directory
}
\description{
function to extract the source directory for a tagstack
}
\seealso{
Other slot access functions: \code{\link{DeployID}},
  \code{\link{Ptt}}, \code{\link{filename}},
  \code{\link{instrument}}, \code{\link{loadtime}},
  \code{\link{location}}, \code{\link{species}},
  \code{\link{streamtype}}, \code{\link{tag_en}},
  \code{\link{tag_st}}, \code{\link{tagdir}}
}
\concept{slot access functions}
