% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torad.R
\name{torad}
\alias{torad}
\title{convert degrees to radians}
\usage{
torad(ang)
}
\arguments{
\item{ang}{angle in degrees}
}
\value{
angle in radians
}
\description{
convert degrees to radians
}
\examples{
torad(180)
}
\seealso{
Other distance functions: \code{\link{latlond.hav}}
}
\concept{distance functions}
