% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wch_html2df.R
\name{wch_html2df}
\alias{wch_html2df}
\title{convert a wch report file (html) to data.frame}
\usage{
wch_html2df(dir, file)
}
\arguments{
\item{dir}{a dirctory of wch report files (html) to convert.}

\item{file}{a single wch report file (html) to convert}
}
\value{
a \code{data.frame}. if multiple input files than each row is a new file. columns correspond to fields.
}
\description{
will convert the fields in a standard mk10 wch report file (html) or a directory of such files generated by mk10 host into a \code{data.frame} with fields as columns and ready for export into a csv.
}
