% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcsv.R
\name{wcsv}
\alias{wcsv}
\title{sensible defaults for \code{write.table}}
\usage{
wcsv(..., sep = ",", row.names = FALSE)
}
\arguments{
\item{...}{all the regular good stuff for finding files and whatnot.}

\item{sep}{we're talking about comma seperated values here.}

\item{row.names}{defaults to FALSE, because that's usually what i want. not when i'm printing a matrix or something though. but here for tables, yes this is what i want.}
}
\description{
sensible defaults for \code{write.table}
}
\details{
please don't use write_csv or any of that other nonesense, it breaks everyone's code...
}
\examples{
# use it just like write.table but without the worry
x <- data.frame(x1 = rnorm(10), x2 = rnorm(10))
wcsv(x, "file.csv")
}
\seealso{
Other sensible csv functions: \code{\link{rcsv}}
}
\concept{sensible csv functions}
