% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_load_tags.R
\name{batch_load_tags}
\alias{batch_load_tags}
\title{batch load sat tags}
\usage{
batch_load_tags(
  data_dir,
  tags_dir = NA,
  ignore = NA,
  streams = NA,
  stream_delim = "-",
  retain_date_format = FALSE
)
}
\arguments{
\item{data_dir}{character points to a directory of sat tags}

\item{tags_dir}{character vector of tag directory names to include or NA defaults to every directory in \code{data_dir}}

\item{ignore}{character vector of directories to ignore or NA defaults not to ignore any directories.}

\item{streams}{character vector limiting which streams to search for. NA defaults to all streams. note \code{*-Summary.csv} and \code{*-Labels.csv} are expected to populate some of the slots of \code{sattag-class}.}

\item{stream_delim}{character defaults to \code{"-"}. this is what the wildlife computers portal puts between the tag identifier (sometimes DeployID, sometimes Ptt) and the stream identifier (e.g., Argos, RTC, etc.) in the csv files.}

\item{retain_date_format}{bool defaults to FALSE. set to TRUE to keep whatever date format the source files are in and not add any additional columns for datenum formatted dates. gets passed to \code{\link[sattagutils]{load_tag}}.}
}
\value{
an S4 object of class \code{\link[sattagutils]{tagstack}}
}
\description{
load sat tags from a directory containing directories downloaded from the wildlife computers portal.
}
\examples{
\dontrun{
tags <- batch_load_tags("path/to/tags/")
}
}
\seealso{
Other tag stream loaders: 
\code{\link{load_tag}()}
}
\concept{tag stream loaders}
