% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream.R, R/sattag.R, R/tagstack.R
\docType{methods}
\name{filename}
\alias{filename}
\alias{filename,sattagstream-method}
\alias{filename,sattag-method}
\alias{filename,tagstack-method}
\title{get filename of a sattagstream}
\usage{
filename(x)

\S4method{filename}{sattagstream}(x)

\S4method{filename}{sattag}(x)

\S4method{filename}{tagstack}(x)
}
\arguments{
\item{x}{source filename for a csv data stream file.}
}
\description{
use this function to get the filename of a \code{\link[sattagutils]{sattagstream-class}}. changing the source filename after construction is currently not supported.
}
\section{Methods (by class)}{
\itemize{
\item \code{sattagstream}: get the source filename of a sattagstream

\item \code{sattag}: return a vector of source filenames of all sattagstreams contained in a sattag

\item \code{tagstack}: return a vector of source filenames of all sattagstreams contained in a tagstack
}}

\seealso{
\code{\link[sattagutils]{sattagstream}}

Other slot access functions: \code{\link{DeployID}},
  \code{\link{Ptt}}, \code{\link{instrument}},
  \code{\link{loadtime}}, \code{\link{location}},
  \code{\link{species}}, \code{\link{streamtype}},
  \code{\link{tag_en}}, \code{\link{tag_st}},
  \code{\link{tagdir}}, \code{\link{tagstackdir}}
}
\concept{slot access functions}
