% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattag.R
\docType{class}
\name{sattag-class}
\alias{sattag-class}
\title{an S4 class to represent a single sat tag}
\description{
represents a single sat tag. holds a small amount of meta data and a list of data streams.
}
\section{Slots}{

\describe{
\item{\code{instrument}}{instrument type, e.g., "MK10-A"}

\item{\code{DeployID}}{character vector (pulled from summary by \code{load_tag})}

\item{\code{Ptt}}{character vector (pulled from summary stream by \code{load_tag})}

\item{\code{species}}{character vector (pulled from summary stream by \code{load_tag})}

\item{\code{location}}{character vector (pulled from summary stream by \code{load_tag})}

\item{\code{earliestdata,latestdata}}{numeric dates (seconds since UNIX epoch). can either by pulled from summary stream by \code{load_tag} or user updated to represent censoring.}

\item{\code{nstreams}}{number of different data streams set automatically by \code{load_tag}}

\item{\code{streams}}{a list of \code{\link[sattagutils]{sattagstream-class}} objects.}

\item{\code{directory}}{the original directory from which the tag was loaded if set by \code{load_tag}}

\item{\code{loadtime}}{a character vector set by \code{initialize()} when object is instantiated.}
}}

