% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream.R
\docType{class}
\name{sattagstream-class}
\alias{sattagstream-class}
\title{an S4 class to represent a single data stream in a tag}
\description{
a parent class for specific sattag stream types. generally accessed from within a \code{\link[sattagutils]{sattag}}.
}
\section{Slots}{

\describe{
\item{\code{streamtype}}{character of the stream type for convenience (e.g., argos, behavior, etc.).}

\item{\code{filename}}{the original filename from which the data stream was derrived. this will almost always be a text or csv file.}
}}

\seealso{
extends \code{\link[sattagutils]{es4dataframe}}.
}
