% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream.R
\docType{class}
\name{sattagstream-class}
\alias{sattagstream-class}
\title{an S4 class to represent a single data stream in a tag}
\description{
an S4 class to represent a single data stream in a tag
}
\section{Slots}{

\describe{
\item{\code{filename}}{the original filename from which the data stream was derrived. this will almost always be a text or csv file.}

\item{\code{data}}{a data frame representing the actual data stream. almost all wildlife computer data streams (csvs) are in tabular format so this makes sense, but in the future perhaps this should be made more general (e.g., list).}
}}

