% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream_show.R
\name{show,sattagstream-method}
\alias{show,sattagstream-method}
\title{display an S4 object which extends \code{sattagstream}}
\usage{
\S4method{show}{sattagstream}(object)
}
\description{
generic method for show.
}
