% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sattagstream.R, R/sattag.R, R/tagstack.R
\docType{methods}
\name{streamtype}
\alias{streamtype}
\alias{streamtype,sattagstream-method}
\alias{streamtype,sattag-method}
\alias{streamtype,tagstack-method}
\title{get stream type of a sattagstream}
\usage{
streamtype(x)

\S4method{streamtype}{sattagstream}(x)

\S4method{streamtype}{sattag}(x)

\S4method{streamtype}{tagstack}(x)
}
\arguments{
\item{x}{stream type.}
}
\description{
use this function to get the stream type of a \code{\link[sattagutils]{sattagstream-class}}. changing the stream type after construction is currently not supported.
}
\section{Methods (by class)}{
\itemize{
\item \code{sattagstream}: get the streamtype of a sattagstream

\item \code{sattag}: return a vector of stream types of all sattagstreams contained in a sattag

\item \code{tagstack}: return a vector of stream types of all sattagstreams contained in a tagstack
}}

\seealso{
\code{\link[sattagutils]{sattagstream}}

Other slot access functions: \code{\link{DeployID}},
  \code{\link{Ptt}}, \code{\link{filename}},
  \code{\link{instrument}}, \code{\link{loadtime}},
  \code{\link{location}}, \code{\link{species}},
  \code{\link{tag_en}}, \code{\link{tag_st}},
  \code{\link{tagdir}}, \code{\link{tagstackdir}}
}
\concept{slot access functions}
